+++ src/odin/enhancedtrippath.cc
@@ -423,7 +423,7 @@ std::vector<std::string> EnhancedTripLeg_Edge::GetLevelRef() const {
 
   // try to get level_refs, else create some from levels as fallback
   if (!tagged_value().empty()) {
-    for (int t = 0; t < tagged_value().size(); ++t) {
+    for (int t = 0; t < (int) tagged_value().size(); ++t) {
       if (tagged_value()[t].type() == TaggedValue_Type_kLevelRef) {
         level_refs.emplace_back(tagged_value()[t].value());
       } else if (tagged_value()[t].type() == TaggedValue_Type_kLevels) {
