+++ test/gurka/test_shortest.cc
@@ -105,7 +105,7 @@ TEST(AutoShorter, deprecation) {
   ParseApi(request_str, Options::route, request);
 
   ASSERT_EQ(request.options().costing_type(), Costing::auto_);
-  const auto& options = request.options().costings().find(Costing::auto_)->second.options();
+  const auto& options = request.options().costings().find((int) Costing::auto_)->second.options();
   ASSERT_EQ(options.shortest(), true);
   ASSERT_EQ(options.use_ferry(), 0.1f);
   ASSERT_EQ(options.use_tolls(), 0.77f);
